reword embed utils
