+++ utils/util.ts
@@ -1,26 +1,5 @@
-import { Client, MessageEmbed } from "discord.js";
 import fs from "node:fs";
 
-export function create_default_embed(
-  client: Client,
-  title: string,
-  description: string
-) {
-  const color = "#0099ff";
-  const thumbnail =
-    "https://playantares.com/resources/CSSC-bot/cssc-server-icon.png";
-  const footer = `Delivered in: ${client.ws.ping}ms | CSSC-bot | ${process.env.VERSION}`;
-  const footerIcon = "https://playantares.com/resources/CSSC-bot/icon.jpg";
-
-  // Embed construction
-  const embed = new MessageEmbed()
-    .setColor(color)
-    .setTitle(title)
-    .setThumbnail(thumbnail)
-    .setDescription(description)
-    .setFooter({ text: footer, iconURL: footerIcon });
-  return embed;
-}
 
 export function sleep(ms: number) {
   // Create new promise that resolves itself after a delay of <ms>
